ROF calculations working, but results are yet to be tested for continuity and ROF values for more complex systems.
There is a lag in the available storage for the 1Ut2Res test that needs to be checked. Also, in the simulation test, the ROF two is getting the storage values in the beginning, which is weird.

+++ src/SystemComponents/WaterSources/Reservoir.cpp
@@ -15,7 +15,7 @@ Reservoir::Reservoir(const string &source_name, const int id, const double min_e
         catchments, online,
         capacity,
         "Reservoir") {
-    storage_records.assign(catchments[0]->series_length, 0);
+    storage_records.assign((unsigned long) catchments[0]->series_length, 0);
 }
 
 /**
